# internal/goa/service/resume/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: resume API
    description: go-goa API with goa framework
    contact:
        name: hiromaily
        url: https://hiromaily.github.io/
        email: hiromaily@gmail.com
    license:
        name: MIT
        url: https://github.com/goadesign/goa/blob/master/LICENSE
    version: "1.0"
servers:
    - url: http://localhost:8080/api
paths:
...
    /company/branch:
        post:
            tags:
                - hy_companybranch
            summary: createCompanyBranch hy_companybranch
            description: Create new company branch
            operationId: hy_companybranch#createCompanyBranch
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCompanyBranchRequestBody'
                        example:
                            address: Shinagawa Tokyo
                            company_id: 2726521388268420732
                            country_id: 110
            responses:
                "200":
                    description: OK response.
                "201":
                    description: Created response.
            security:
                - jwt_header_Authorization:
                    - api:access
    /company/branch/{company_branch_id}:
        delete:
            tags:
                - hy_companybranch
            summary: deleteCompanyBranch hy_companybranch
            description: Delete company branch
            operationId: hy_companybranch#deleteCompanyBranch
            parameters:
                - name: company_branch_id
                  in: path
                  description: Company branch ID
                  required: true
                  schema:
                    type: integer
                    description: Company branch ID
                    example: 2412966896863543046
                    format: int64
                  example: 8964776204148672013
            responses:
                "200":
                    description: OK response.
            security:
                - jwt_header_Authorization:
                    - api:access
        get:
            tags:
                - hy_companybranch
            summary: getCompanyBranch hy_companybranch
            description: Get company branch with given id
            operationId: hy_companybranch#getCompanyBranch
            parameters:
                - name: company_branch_id
                  in: path
                  description: Company branch ID
                  required: true
                  schema:
                    type: integer
                    description: Company branch ID
                    example: 2764320691338423418
                    format: int64
                  example: 5156853143016840597
            responses:
                "200":
                    description: OK response.
            security:
                - jwt_header_Authorization:
                    - api:access
        put:
            tags:
                - hy_companybranch
            summary: updateCompanyBranch hy_companybranch
            description: Change company branch properties
            operationId: hy_companybranch#updateCompanyBranch
            parameters:
                - name: company_branch_id
                  in: path
                  description: Company branch ID
                  required: true
                  schema:
                    type: integer
                    description: Company branch ID
                    example: 2183358531700565598
                    format: int64
                  example: 1590406795724218417
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateCompanyBranchRequestBody'
                        example:
                            address: Shinagawa Tokyo
                            country_id: 110
            responses:
                "200":
                    description: OK response.
            security:
                - jwt_header_Authorization:
                    - api:access
...
components:
    schemas:
...
        CreateCompanyBranchRequestBody:
            type: object
            properties:
                address:
                    type: string
                    description: Company Address
                    example: Shinagawa Tokyo
                    minLength: 2
                    maxLength: 80
                company_id:
                    type: integer
                    description: Company ID
                    example: 8029950485983377782
                    format: int64
                country_id:
                    type: integer
                    description: Country ID
                    example: 110
                    minimum: 1
                    maximum: 999
            example:
                address: Shinagawa Tokyo
                company_id: 4639867356943733610
                country_id: 110
            required:
                - country_id
                - address
...
        UpdateCompanyBranchRequestBody:
            type: object
            properties:
                address:
                    type: string
                    description: Company Address
                    example: Shinagawa Tokyo
                    minLength: 2
                    maxLength: 80
                country_id:
                    type: integer
                    description: Country ID
                    example: 110
                    minimum: 1
                    maximum: 999
            example:
                address: Shinagawa Tokyo
                country_id: 110
            required:
                - country_id
                - address
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            scheme: bearer
...